Make function layer toggle-able
For the moment, the function layer is toggle-able only from the base
layer.  From the key-pad layer, it acts as a momentary switch. As I do
not fully understand how layers interact, and I do not plan to use the
key-pad layer, I think this is reasonable for the moment.